+++ common/src/utils/array-utils/array-utils.ts
@@ -1,4 +1,3 @@
-// TODO: move out of dbutils
 export function chunkArray<T>(array: Array<T>, chunkSize: number): Iterable<Array<T>> {
   const shallowCopy = [...array];